category attribute added

+++ inventory_service/src/main/java/com/GigaGrocery/product_service/dto/StockRequest.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,5 +14,5 @@ import java.time.LocalDateTime;
 public class StockRequest {
     private LocalDateTime date;
     private int quantity;
-    private BigDecimal price;
+    private float price;
 }
