+++ inventory_service/src/main/java/com/GigaGrocery/product_service/dto/StockResponse.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,5 +15,5 @@ import java.time.LocalDateTime;
 public class StockResponse {
     private LocalDateTime date;
     private int quantity;
-    private BigDecimal price;
+    private float price;
 }
